requirements.txt
openai==0.19.0
gTTS==2.3.1
mysql-connector-python==8.0.32
google-api-python-client==2.80.0
